refactor: remove monorepo OIDC UserInfo augmentation, consume CSL 0.1.0-alpha31
CSL 0.1.0-alpha31 ships its own UserInfo augmentation via
OidcClaimsProviderConfiguration, making the parallel monorepo implementations
redundant. Bump the CSL version and remove the duplicated code.

- Delete OidcClaimsProvider, CachingOidcClaimsProvider, NoopOidcClaimsProvider,
  OidcUserInfoClient, OidcUserInfoException from security-core (plus their tests)
- Remove oidcClaimsProvider() bean from OidcOverrideBeansConfiguration; CSL's
  OidcClaimsProviderConfiguration now owns this bean via @ConditionalOnMissingBean
- Retain oidcUserInfoHttpClient() bean so SSL bundle support carries over
- Update all OidcClaimsProvider imports from io.camunda.security.oidc to
  io.camunda.security.api.context across zeebe/broker, zeebe/gateway-grpc, dist
- Replace new NoopOidcClaimsProvider() with lambda (jwtClaims, tokenValue) ->
  jwtClaims in non-spring modules that lack the spring-boot-starter dependency

Closes #406 (camunda/camunda-security-library)
