--- security/security-core/src/test/java/io/camunda/security/oidc/NoopOidcClaimsProviderTest.java
@@ -1,26 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.security.oidc;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.Map;
-import org.junit.jupiter.api.Test;
-
-class NoopOidcClaimsProviderTest {
-
-  @Test
-  void returnsJwtClaimsUnchanged() {
-    final var provider = new NoopOidcClaimsProvider();
-    final var input = Map.<String, Object>of("sub", "alice", "iss", "https://idp.example");
-
-    final var output = provider.claimsFor(input, "token-abc");
-
-    assertThat(output).isSameAs(input);
-  }
-}
